Move storagehotplug validation to pkg/storage
