+++ pkg/storage/admitters/BUILD.bazel
@@ -5,6 +5,7 @@ go_test(
     srcs = [
         "admit_suite_test.go",
         "disks_test.go",
+        "storagehotplug_test.go",
         "vm-storage-admitter_test.go",
         "vmexport_test.go",
         "vmrestore_test.go",
@@ -15,8 +16,10 @@ go_test(
         "//pkg/apimachinery/patch:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/testutils:go_default_library",
+        "//pkg/util/webhooks:go_default_library",
         "//pkg/virt-api/webhooks:go_default_library",
         "//pkg/virt-config:go_default_library",
+        "//pkg/virt-config/featuregate:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/export/v1beta1:go_default_library",
         "//staging/src/kubevirt.io/api/snapshot/v1beta1:go_default_library",
@@ -29,6 +32,7 @@ go_test(
         "//vendor/go.uber.org/mock/gomock:go_default_library",
         "//vendor/k8s.io/api/admission/v1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
@@ -45,6 +49,7 @@ go_library(
     srcs = [
         "data-volume-template.go",
         "disks.go",
+        "storagehotplug.go",
         "vm-storage-admitter.go",
         "vm-storage-status.go",
         "vmexport.go",
