+++ pkg/storage/admitters/disks.go
@@ -381,61 +381,3 @@ func ValidatePath(field *k8sfield.Path, path string) []metav1.StatusCause {
 
 	return causes
 }
-
-func ValidateHotplugDiskConfiguration(disk *v1.Disk, name, messagePrefix, field string) []metav1.StatusCause {
-	if disk == nil {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: fmt.Sprintf("%s for [%s] requires the disk field to be set.", messagePrefix, name),
-			Field:   field,
-		}}
-	}
-
-	bus := getDiskBus(*disk)
-	switch {
-	case disk.DiskDevice.Disk != nil:
-		if bus != v1.DiskBusSCSI && bus != v1.DiskBusVirtio {
-			return []metav1.StatusCause{{
-				Type:    metav1.CauseTypeFieldValueInvalid,
-				Message: fmt.Sprintf("%s for disk [%s] requires bus to be 'scsi' or 'virtio'. [%s] is not permitted.", messagePrefix, name, bus),
-				Field:   field,
-			}}
-		}
-	case disk.DiskDevice.LUN != nil:
-		if bus != v1.DiskBusSCSI {
-			return []metav1.StatusCause{{
-				Type:    metav1.CauseTypeFieldValueInvalid,
-				Message: fmt.Sprintf("%s for LUN [%s] requires bus to be 'scsi'. [%s] is not permitted.", messagePrefix, name, bus),
-				Field:   field,
-			}}
-		}
-	default:
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: fmt.Sprintf("%s for [%s] requires diskDevice of type 'disk' or 'lun' to be used.", messagePrefix, name),
-			Field:   field,
-		}}
-	}
-
-	if disk.DedicatedIOThread != nil && *disk.DedicatedIOThread && bus != v1.DiskBusVirtio {
-		return []metav1.StatusCause{{
-			Type:    metav1.CauseTypeFieldValueInvalid,
-			Message: fmt.Sprintf("%s for [%s] requires virtio bus for IOThreads.", messagePrefix, name),
-			Field:   field,
-		}}
-	}
-
-	// Validate boot order
-	if disk.BootOrder != nil {
-		order := *disk.BootOrder
-		if order < 1 {
-			return []metav1.StatusCause{{
-				Type:    metav1.CauseTypeFieldValueInvalid,
-				Message: fmt.Sprintf("spec.domain.devices.disks[1] must have a boot order > 0, if supplied"),
-				Field:   fmt.Sprintf("spec.domain.devices.disks[1].bootOrder"),
-			}}
-		}
-	}
-
-	return nil
-}
