+++ pkg/virt-api/webhooks/validating-webhook/admitters/vmi-update-admitter.go
@@ -21,7 +21,6 @@ package admitters
 
 import (
 	"context"
-	"fmt"
 	"strings"
 
 	admissionv1 "k8s.io/api/admission/v1"
@@ -130,233 +129,6 @@ func (admitter *VMIUpdateAdmitter) Admit(_ context.Context, ar *admissionv1.Admi
 	}
 }
 
-func validateExpectedDisksAndFilesystems(volumes []v1.Volume, disks []v1.Disk, filesystems []v1.Filesystem, config *virtconfig.ClusterConfig) error {
-	names := make(map[string]struct{})
-	for _, volume := range volumes {
-		if volume.MemoryDump == nil {
-			names[volume.Name] = struct{}{}
-		}
-	}
-
-	requiredVolumes := len(filesystems)
-	for _, disk := range disks {
-		_, ok := names[disk.Name]
-		// it's okay for a CDRom to not be mapped to a volume
-		if !ok && disk.CDRom != nil && config.DeclarativeHotplugVolumesEnabled() {
-			continue
-		}
-		requiredVolumes++
-	}
-
-	// Make sure volume is not mapped to multiple disks/filesystems or volume mapped to nothing
-	if requiredVolumes != len(names) {
-		return fmt.Errorf("mismatch between volumes declared (%d) and required (%d)", len(names), requiredVolumes)
-	}
-
-	return nil
-}
-
-// admitStorageUpdate compares the old and new volumes and disks, and ensures that they match and are valid.
-func admitStorageUpdate(newVolumes, oldVolumes []v1.Volume, newDisks, oldDisks []v1.Disk, volumeStatuses []v1.VolumeStatus, newVMI *v1.VirtualMachineInstance, config *virtconfig.ClusterConfig) *admissionv1.AdmissionResponse {
-	if err := validateExpectedDisksAndFilesystems(newVolumes, newDisks, newVMI.Spec.Domain.Devices.Filesystems, config); err != nil {
-		return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-			{
-				Type:    metav1.CauseTypeFieldValueInvalid,
-				Message: err.Error(),
-			},
-		})
-	}
-	newHotplugVolumeMap := getHotplugVolumes(newVolumes, volumeStatuses)
-	newPermanentVolumeMap := getPermanentVolumes(newVolumes, volumeStatuses)
-	oldHotplugVolumeMap := getHotplugVolumes(oldVolumes, volumeStatuses)
-	oldPermanentVolumeMap := getPermanentVolumes(oldVolumes, volumeStatuses)
-	migratedVolumeMap := getMigratedVolumeMaps(newVMI.Status.MigratedVolumes)
-
-	newDiskMap := getDiskMap(newDisks)
-	oldDiskMap := getDiskMap(oldDisks)
-
-	permanentAr := verifyPermanentVolumes(newPermanentVolumeMap, oldPermanentVolumeMap, newDiskMap, oldDiskMap, migratedVolumeMap)
-	if permanentAr != nil {
-		return permanentAr
-	}
-
-	hotplugAr := verifyHotplugVolumes(newHotplugVolumeMap, oldHotplugVolumeMap, newDiskMap, oldDiskMap, migratedVolumeMap)
-	if hotplugAr != nil {
-		return hotplugAr
-	}
-
-	return nil
-}
-
-func verifyHotplugVolumes(newHotplugVolumeMap, oldHotplugVolumeMap map[string]v1.Volume, newDisks, oldDisks map[string]v1.Disk,
-	migratedVols map[string]bool) *admissionv1.AdmissionResponse {
-	for k, v := range newHotplugVolumeMap {
-		if _, ok := oldHotplugVolumeMap[k]; ok {
-			_, okMigVol := migratedVols[k]
-			// New and old have same volume, ensure they are the same
-			if !equality.Semantic.DeepEqual(v, oldHotplugVolumeMap[k]) && !okMigVol {
-				return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-					{
-						Type:    metav1.CauseTypeFieldValueInvalid,
-						Message: fmt.Sprintf("hotplug volume %s, changed", k),
-					},
-				})
-			}
-			if v.MemoryDump == nil {
-				if _, ok := newDisks[k]; !ok {
-					return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-						{
-							Type:    metav1.CauseTypeFieldValueInvalid,
-							Message: fmt.Sprintf("volume %s doesn't have a matching disk", k),
-						},
-					})
-				}
-				if !equality.Semantic.DeepEqual(newDisks[k], oldDisks[k]) {
-					return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-						{
-							Type:    metav1.CauseTypeFieldValueInvalid,
-							Message: fmt.Sprintf("hotplug disk %s, changed", k),
-						},
-					})
-				}
-			}
-		} else {
-			// This is a new volume, ensure that the volume is either DV, PVC or memoryDumpVolume
-			if v.DataVolume == nil && v.PersistentVolumeClaim == nil && v.MemoryDump == nil {
-				return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-					{
-						Type:    metav1.CauseTypeFieldValueInvalid,
-						Message: fmt.Sprintf("volume %s is not a PVC or DataVolume", k),
-					},
-				})
-			}
-			if v.MemoryDump == nil {
-				// Also ensure the matching new disk exists and has a valid bus
-				if _, ok := newDisks[k]; !ok {
-					return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-						{
-							Type:    metav1.CauseTypeFieldValueInvalid,
-							Message: fmt.Sprintf("disk %s does not exist", k),
-						},
-					})
-				}
-				disk := newDisks[k]
-				if _, ok := oldDisks[k]; !ok {
-					causes := storageadmitters.ValidateHotplugDiskConfiguration(&disk, k, "Hotplug configuration", "")
-					if len(causes) > 0 {
-						return webhookutils.ToAdmissionResponse(causes)
-					}
-				}
-			}
-		}
-	}
-	return nil
-}
-
-func isMigratedVolume(newVol, oldVol *v1.Volume, migratedVolumeMap map[string]bool) bool {
-	if newVol.Name != oldVol.Name {
-		return false
-	}
-	_, ok := migratedVolumeMap[newVol.Name]
-	return ok
-}
-
-func verifyPermanentVolumes(newPermanentVolumeMap, oldPermanentVolumeMap map[string]v1.Volume, newDisks, oldDisks map[string]v1.Disk, migratedVolumeMap map[string]bool) *admissionv1.AdmissionResponse {
-	if len(newPermanentVolumeMap) != len(oldPermanentVolumeMap) {
-		// Removed one of the permanent volumes, reject admission.
-		return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-			{
-				Type:    metav1.CauseTypeFieldValueInvalid,
-				Message: "Number of permanent volumes has changed",
-			},
-		})
-	}
-
-	// Ensure we didn't modify any permanent volumes
-	for k, v := range newPermanentVolumeMap {
-		// Know at this point the new old and permanent have the same count.
-		if _, ok := oldPermanentVolumeMap[k]; !ok {
-			return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-				{
-					Type:    metav1.CauseTypeFieldValueInvalid,
-					Message: fmt.Sprintf("permanent volume %s, not found", k),
-				},
-			})
-		}
-		oldVol := oldPermanentVolumeMap[k]
-		if isMigratedVolume(&v, &oldVol, migratedVolumeMap) {
-			continue
-		}
-		if !equality.Semantic.DeepEqual(v, oldVol) {
-			return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-				{
-					Type:    metav1.CauseTypeFieldValueInvalid,
-					Message: fmt.Sprintf("permanent volume %s, changed", k),
-				},
-			})
-		}
-		if !equality.Semantic.DeepEqual(newDisks[k], oldDisks[k]) {
-			return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
-				{
-					Type:    metav1.CauseTypeFieldValueInvalid,
-					Message: fmt.Sprintf("permanent disk %s, changed", k),
-				},
-			})
-		}
-	}
-	return nil
-}
-
-func getDiskMap(disks []v1.Disk) map[string]v1.Disk {
-	newDiskMap := make(map[string]v1.Disk, 0)
-	for _, disk := range disks {
-		if disk.Name != "" {
-			newDiskMap[disk.Name] = disk
-		}
-	}
-	return newDiskMap
-}
-
-func getHotplugVolumes(volumes []v1.Volume, volumeStatuses []v1.VolumeStatus) map[string]v1.Volume {
-	permanentVolumesFromStatus := make(map[string]v1.Volume, 0)
-	for _, volume := range volumeStatuses {
-		if volume.HotplugVolume == nil {
-			permanentVolumesFromStatus[volume.Name] = v1.Volume{}
-		}
-	}
-	permanentVolumes := make(map[string]v1.Volume, 0)
-	for _, volume := range volumes {
-		if _, ok := permanentVolumesFromStatus[volume.Name]; !ok {
-			permanentVolumes[volume.Name] = volume
-		}
-	}
-	return permanentVolumes
-}
-
-func getPermanentVolumes(volumes []v1.Volume, volumeStatuses []v1.VolumeStatus) map[string]v1.Volume {
-	permanentVolumesFromStatus := make(map[string]v1.Volume, 0)
-	for _, volume := range volumeStatuses {
-		if volume.HotplugVolume == nil {
-			permanentVolumesFromStatus[volume.Name] = v1.Volume{}
-		}
-	}
-	permanentVolumes := make(map[string]v1.Volume, 0)
-	for _, volume := range volumes {
-		if _, ok := permanentVolumesFromStatus[volume.Name]; ok {
-			permanentVolumes[volume.Name] = volume
-		}
-	}
-	return permanentVolumes
-}
-
-func getMigratedVolumeMaps(migratedDisks []v1.StorageMigratedVolumeInfo) map[string]bool {
-	volumes := make(map[string]bool)
-	for _, v := range migratedDisks {
-		volumes[v.VolumeName] = true
-	}
-	return volumes
-}
-
 func admitVMILabelsUpdate(
 	newVMI *v1.VirtualMachineInstance,
 	oldVMI *v1.VirtualMachineInstance,
@@ -404,7 +176,7 @@ func admitHotplug(
 		return response
 	}
 
-	return admitStorageUpdate(
+	return storageadmitters.AdmitHotplugStorage(
 		newVMI.Spec.Volumes,
 		oldVMI.Spec.Volumes,
 		newVMI.Spec.Domain.Devices.Disks,
